requirements.txt
googleapiclient
absl-py==0.9.0
adal==1.2.4
astor==0.8.1
azure-cognitiveservices-search-newssearch==2.0.0
azure-common==1.1.25
beautifulsoup4==4.9.1
boto3==1.12.35
botocore==1.15.35
bs4==0.0.1
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
click==7.1.1
cryptography==3.1
cssselect==1.1.0
docutils==0.15.2
feedfinder2==0.0.4
feedparser==6.0.1
filelock==3.0.12
gast==0.2.2
google-api-core==1.22.2
google-api-python-client==1.12.2
google-auth==1.21.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-language==1.3.0
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
html2text==2020.1.16
httplib2==0.18.1
ibm-cloud-sdk-core==1.7.3
ibm-watson==4.7.1
idna==2.9
isodate==0.6.0
jieba3k==0.35.1
jmespath==0.9.5
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lxml==4.5.2
Markdown==3.2.1
msrest==0.6.19
msrestazure==0.6.4
nltk==3.5
numpy==1.18.2
oauthlib==3.1.0
opt-einsum==3.2.0
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.2.20
requests==2.23.0
requests-file==1.5.1
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
sacremoses==0.0.38
scipy==1.4.1
sentencepiece==0.1.85
sgmllib3k==1.0.0
six==1.14.0
soupsieve==2.0.1
termcolor==1.1.0
tinysegmenter==0.3
tldextract==2.2.3
tokenizers==0.5.2
tqdm==4.45.0
transformers==2.7.0
uritemplate==3.0.1
urllib3==1.25.8
websocket-client==0.48.0
Werkzeug==1.0.1
wrapt==1.12.1
